Issue  #3 BasicHomePage Take1
Did a LOT in Home.razor. I think TemplatedBar was just shit I was working on, so you can delete it if you think its dumb. Uh, I couldn't figure out how to rename/delete pages! So If I could maybe be explained that because well, you'll see.

DOWN TO HOME CONTENT>>>

Div CreateHouse is the entire "Add A New Home" Button Class.

The JoinButton will eventually also have a Div JoinHouse but im all about placeholders so no.

The Code Section, I think I might have code that I don't need, but I for the life of me can not figure out where anymore. I'm down for new ideas on whaat to call things! I definitely got myself a little confused...

Alsoo, the "Click to Create a GiftHouse Button" is also a placeholder. It does nothing  atm. The gameplan is for it to open up into the "createHouse"div

This is barebones, to be continued after its checked out.