Images/Components
Images should be fine now, let me know if they aren't.

Made the Bottom Nav it's own component, and I want to ask about implemeting code thats conditional for if they're on the homepage or not, so that if theyre not it adds a back button to go back to the homepage. (Not sure if this makes sense but I can clarify another time)